requirements.txt
apache-airflow-providers-docker
apache-airflow-providers-apache-spark
docker
kafka-python